refactor(lib): Remove experimental Next implementation
