+++ android/src/fabric/java/com/reactnativepagerview/PagerViewViewManager.kt
@@ -173,11 +173,6 @@ class PagerViewViewManager : ViewGroupManager<NestedScrollableHost>(), RNCViewPa
         return
     }
 
-    @ReactProp(name = "useLegacy")
-    override fun setUseLegacy(view: NestedScrollableHost?, value: Boolean) {
-        return
-    }
-
     fun goTo(root: NestedScrollableHost?, selectedPage: Int, scrollWithAnimation: Boolean) {
         if (root == null) {
             return
