--- android/src/main/java/com/reactnativepagerview/LEGACY_PagerViewViewManagerImpl.kt
@@ -1,7 +1,0 @@
-package com.reactnativepagerview
-
-// Note: The LEGACY_ variant is an iOS-only feature and the related Android files
-// are only included because of and relevant to auxiliary processes, like Codegen.
-object LEGACY_PagerViewViewManagerImpl {
-    const val NAME = "LEGACY_RNCViewPager"
-}
