--- ios/LEGACY/LEGACY_RNCPagerViewManager.h
@@ -1,13 +1,0 @@
-
-#import <UIKit/UIKit.h>
-#import <React/RCTViewManager.h>
-#import <React/RCTUIManager.h>
-#import <React/RCTLog.h>
-#import "LEGACY_RNCPagerView.h"
-NS_ASSUME_NONNULL_BEGIN
-
-@interface LEGACY_RNCPagerViewManager : RCTViewManager
-
-@end
-
-NS_ASSUME_NONNULL_END
