--- ios/RNCPagerScrollView.h
@@ -1,9 +1,0 @@
-/*
-NOTE: This file is *not* currently used, for context see:
-- https://github.com/callstack/react-native-pager-view/pull/783#discussion_r1410295171
-- https://github.com/callstack/react-native-pager-view/pull/783#discussion_r1410316201
-*/
-#include <UIKit/UIKit.h>
-
-@interface RNCPagerScrollView : UIScrollView <UIGestureRecognizerDelegate>
-@end
