+++ ios/RNCPagerViewManager.m
@@ -8,12 +8,14 @@
 RCT_EXPORT_MODULE(RNCViewPager)
 
 RCT_EXPORT_VIEW_PROPERTY(initialPage, NSInteger)
-RCT_EXPORT_VIEW_PROPERTY(orientation, NSString)
+RCT_EXPORT_VIEW_PROPERTY(pageMargin, NSInteger)
 
+RCT_EXPORT_VIEW_PROPERTY(orientation, UIPageViewControllerNavigationOrientation)
 RCT_EXPORT_VIEW_PROPERTY(onPageSelected, RCTDirectEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(onPageScroll, RCTDirectEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(onPageScrollStateChanged, RCTDirectEventBlock)
 RCT_EXPORT_VIEW_PROPERTY(overdrag, BOOL)
+RCT_EXPORT_VIEW_PROPERTY(layoutDirection, NSString)
 
 
 - (void) goToPage
@@ -25,7 +27,7 @@ RCT_EXPORT_VIEW_PROPERTY(overdrag, BOOL)
                                         NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RNCPagerView *view = (RNCPagerView *)viewRegistry[reactTag];
         if (!view || ![view isKindOfClass:[RNCPagerView class]]) {
-            RCTLogError(@"Cannot find ReactNativePageView with tag #%@", reactTag);
+            RCTLogError(@"Cannot find RNCPagerView with tag #%@", reactTag);
             return;
         }
         if (!animated || !view.animating) {
@@ -42,7 +44,7 @@ RCT_EXPORT_VIEW_PROPERTY(overdrag, BOOL)
                                         NSDictionary<NSNumber *, UIView *> *viewRegistry) {
         RNCPagerView *view = (RNCPagerView *)viewRegistry[reactTag];
         if (!view || ![view isKindOfClass:[RNCPagerView class]]) {
-            RCTLogError(@"Cannot find ReactNativePageView with tag #%@", reactTag);
+            RCTLogError(@"Cannot find RNCPagerView with tag #%@", reactTag);
             return;
         }
         [view shouldScroll:enabled];
@@ -61,7 +63,7 @@ RCT_EXPORT_METHOD(setPageWithoutAnimation
     [self goToPage:reactTag index:index animated:false];
 }
 
-RCT_EXPORT_METHOD(setScrollEnabled
+RCT_EXPORT_METHOD(setScrollEnabledImperatively
                   : (nonnull NSNumber *)reactTag enabled
                   : (nonnull NSNumber *)enabled) {
     BOOL isEnabled = [enabled boolValue];
