+++ src/PagerView.tsx
@@ -1,10 +1,9 @@
 import React from 'react';
-import { Platform, Keyboard, StyleProp, ViewStyle } from 'react-native';
+import { Platform, Keyboard } from 'react-native';
 import { I18nManager } from 'react-native';
 import type * as ReactNative from 'react-native';
 
 import {
-  LEGACY_childrenWithOverriddenStyle,
   childrenWithOverriddenStyle,
 } from './utils';
 
@@ -14,11 +13,8 @@ import PagerViewNativeComponent, {
   OnPageScrollStateChangedEventData,
   OnPageSelectedEventData,
   NativeProps,
-} from './specs/PagerViewNativeComponent';
+} from './PagerViewNativeComponent';
 
-import LEGACY_PagerViewNativeComponent, {
-  Commands as LEGACY_PagerViewNativeCommands,
-} from './specs/LEGACY_PagerViewNativeComponent';
 
 /**
  * Container that allows to flip left and right between child views. Each
@@ -62,27 +58,10 @@ import LEGACY_PagerViewNativeComponent, {
  * ```
  */
 
-// Type that excludes the private methods
-// Fix for class exporting using hook.
-type PublicPagerViewMethods = Omit<
-  PagerViewInternal,
-  | 'nativeCommandsWrapper'
-  | 'deducedLayoutDirection'
-  | '_onPageScroll'
-  | '_onPageScrollStateChanged'
-  | '_onPageSelected'
-  | '_onMoveShouldSetResponderCapture'
->;
-
-class PagerViewInternal extends React.Component<NativeProps> {
+export class PagerView extends React.Component<NativeProps> {
   private isScrolling = false;
   pagerView: React.ElementRef<typeof PagerViewNativeComponent> | null = null;
 
-  private get nativeCommandsWrapper() {
-    return this.props.useLegacy
-      ? LEGACY_PagerViewNativeCommands
-      : PagerViewNativeCommands;
-  }
 
   private get deducedLayoutDirection() {
     if (
@@ -138,7 +117,7 @@ class PagerViewInternal extends React.Component<NativeProps> {
    */
   public setPage = (selectedPage: number) => {
     if (this.pagerView) {
-      this.nativeCommandsWrapper.setPage(this.pagerView, selectedPage);
+      PagerViewNativeCommands.setPage(this.pagerView, selectedPage);
     }
   };
 
@@ -148,7 +127,7 @@ class PagerViewInternal extends React.Component<NativeProps> {
    */
   public setPageWithoutAnimation = (selectedPage: number) => {
     if (this.pagerView) {
-      this.nativeCommandsWrapper.setPageWithoutAnimation(
+      PagerViewNativeCommands.setPageWithoutAnimation(
         this.pagerView,
         selectedPage
       );
@@ -162,7 +141,7 @@ class PagerViewInternal extends React.Component<NativeProps> {
    */
   public setScrollEnabled = (scrollEnabled: boolean) => {
     if (this.pagerView) {
-      this.nativeCommandsWrapper.setScrollEnabledImperatively(
+      PagerViewNativeCommands.setScrollEnabledImperatively(
         this.pagerView,
         scrollEnabled
       );
@@ -170,10 +149,8 @@ class PagerViewInternal extends React.Component<NativeProps> {
   };
 
   render() {
-    // old iOS `UIPageViewController`-based implementation
-    if (Platform.OS === 'ios' && this.props.useLegacy) {
       return (
-        <LEGACY_PagerViewNativeComponent
+        <PagerViewNativeComponent
           {...this.props}
           ref={(ref) => {
             this.pagerView = ref;
@@ -186,64 +163,8 @@ class PagerViewInternal extends React.Component<NativeProps> {
           onMoveShouldSetResponderCapture={
             this._onMoveShouldSetResponderCapture
           }
-          children={LEGACY_childrenWithOverriddenStyle(this.props.children)}
+          children={childrenWithOverriddenStyle(this.props.children)}
         />
       );
-    }
-
-    const style: StyleProp<ViewStyle> = [
-      this.props.style,
-      this.props.pageMargin
-        ? {
-            marginHorizontal: -this.props.pageMargin / 2,
-          }
-        : null,
-      {
-        flexDirection: this.props.orientation === 'vertical' ? 'column' : 'row',
-      },
-    ];
-
-    // new iOS `UIScrollView`-based implementation, Android, and other platforms
-    return (
-      <PagerViewNativeComponent
-        {...this.props}
-        ref={(ref) => {
-          this.pagerView = ref;
-        }}
-        style={style}
-        layoutDirection={this.deducedLayoutDirection}
-        onPageScroll={this._onPageScroll}
-        onPageScrollStateChanged={this._onPageScrollStateChanged}
-        onPageSelected={this._onPageSelected}
-        onMoveShouldSetResponderCapture={this._onMoveShouldSetResponderCapture}
-        children={childrenWithOverriddenStyle(
-          this.props.children,
-          this.props.pageMargin
-        )}
-      />
-    );
   }
 }
-
-// Temporary solution. It should be removed once all things get fixed
-type PagerViewProps = Omit<NativeProps, 'useLegacy'> & { useNext?: boolean };
-
-export const PagerView = React.forwardRef<
-  PublicPagerViewMethods,
-  PagerViewProps
->((props, ref) => {
-  const { useNext, ...rest } = props;
-  return (
-    <PagerViewInternal
-      {...rest}
-      useLegacy={!useNext}
-      ref={ref as React.LegacyRef<PagerViewInternal>}
-    />
-  );
-});
-
-// React.forwardRef does not type returned component properly, thus breaking Ref<MyComponent> typing.
-// One way to overcome this is using separate typing for component "interface",
-// but that breaks backward compatibility in this case.
-// Approach of merging type is hacky, but produces a good typing for both ref attributes and component itself.
-export type PagerView = PublicPagerViewMethods & typeof PagerView;
