+++ src/index.tsx
@@ -8,7 +8,7 @@ import type {
   OnPageSelectedEventData as PagerViewOnPageSelectedEventData,
   OnPageScrollStateChangedEventData as PageScrollStateChangedNativeEventData,
   NativeProps,
-} from './specs/PagerViewNativeComponent';
+} from './PagerViewNativeComponent';
 
 export type {
   PagerViewOnPageScrollEventData,
